Prevent mysql::server::restart from being set
If mysql is allowed to restart services, on the node
that is responsible for bootstrapping this will cause
issues, since a 'service mysql restart' will take the
cluster down, but will not bring it back up.

The parameter is left as an option, as it may be
desirable in some circumstances.

Note that this does mean that the galera cluster will
not pick up config changes automatically, as config
changes will not trigger a restart of services.