Allow overriding of available_when_readonly
There are cases when you need to override the default puppet setting for
available_when_readonly, such as when doing a rolling upgrade of mysql.
This allows the value to be set. We intentionally use an environment
variable rather than a positional one to avoid having to set the other
vars.