Require status be in place before mysql::server::end
This is needed because a load balancer is probably already looking for
this servicexs to come up.  If other services are waiting on
mysql::server::end and then access MySQL through the load balancer,
then the other parts of the status service may not be available yet.
This ensures that all of them will be before mysql::server::end.