Change debian.cnf after stopping the service
There exists a race where debian.cnf gets changed on a non-master node
before the service is stopped post install of the galera package.  The
end result is that the service stop fails due to the password changing.
This change also fixes some puppet lint errors in the affected file.