Move boostrap check from nc to nmap
nc on RedHat based platforms no longer has the -z option,
so move to nmap instead. nmap was chosen over tcping since
I don't think tcping is in the standard base repos for
Centos, while nmap is in base, and also included in many
default installs.

This change will also affect Debian, but the result should only
be that the nmap package is installed - everything else should work
as before.

A nice side effect is the checking command is also a lot simpler,
since nmap accepts a list as input.