Update to use rspec-puppet facts
This change updates the tests for puppet-galera to use
puppet-rspec-facts for operating system fact management in the tests. By
using puppet-rspec-facts, we get a complete set of facts that fix many
of the previous test failures because facts were missing. By default
puppet-rspec-facts will use the versions from metadata.json for the
operating systes to test.