Add create_root_my_cnf flag
This change adds the ability to disable the management of the root
.my.cnf by the galera and mysql::server classes. This is useful if you
wish to manage your root .my.cnf elsewhere with custom settings. For
example if you wish to use the host parameter within the .my.cnf,
neither the galera or mysql::server classes support this. So you would
need to set create_root_my_cnf to false and manage the file prior to
pulling in the galera class.