Increased the xinetd instances count
The instances count was not set. On centos this defaulted to 50. This is
too small to handle the load of all mysqlchecks for our purposes. Now
it's set to the same value as cps.

This required a newer module of xinetd. The upstream puppet modules had
changed their defaults when it comes to logging. Added old default values
to log failures to the service check.