Install xinetd after packages (#68)
Install the xinetd service file for the status service after the server
package is installed.  The percona packages also include a mysqlchk
xinetd service, and without this change, the puppet module service file
gets installed first, then the package sets it aside.  If xinetd purging
is turned on, then the next puppet run will remove this.  Unfortunately
this means the first puppet run is not idempotent.