Accommodate upstream ordering changes
::mysql::server::config is now before install, so a bunch of things
needed to be moved around to prevent dependency cycles. In addition,
since galera generally won't start without being bootstrapped, debians
require some additional effort to have the default config applied
when the package installs (and starts the service by default) but
the real config applied when the service is started by the module.