Allow Tinkering with mysqlchk Logging
This aims to allow the reduction in verbosity from the mysqlchk xinetd process
for those of us who use it as a liveness check for haproxy.  I've also made the
check user a system account to avoid clashing with uid/gid >= 1000 which are
hard coded in some environments, altering this will not affect existing systems.

In order to get rspec tests working I've had to modernize the API slightly and
move `galera::status` parameters into the top level class.  As a result this
will need a new major revision as it's potentially a breaking change.