Add missing default params for MariaDB other than version 5.5
On Debian 9 the params lookup results in "Function lookup() did not find a value for the name 'galera::mariadb::client_package_name'" due to using a different version than 5.5. The vendor default according to common.yaml is 10.3, but there aren't any params for 'galera::mariadb::103::client_package_name' either.
The actual current version in Debian stable is 10.1. I personally use the MariaDB upstream repo with version 10.3 which I add in my own module and set 'configure_repo' to false. To avoid version specific errors and hundreds of hiera lines for each version this module should use the metapackages instead which always point to the latest version available to APT.