convert to PDK, make test suite not failing
work on #108 and #114

Make the testsuite run again and not fail.
For this a lot of test code was disabled. Since the failing tests were ignored for such a long time, there is a lot to do here.
Disabled test code is marked with `# FIXME`. I removed a bunch of supported OSes from metadata.json. One reason being missing facst in facterdb,
another reason is that those OS-releases might simply not be be really supported. Thats at least the impression I get from the tests.
Also i don't think one needs to list RedHat and CentOS and Scintific. The latter two should be included in simply Redhat.