rework and simplify resource relationships
Previously, both the galera::status and the galera::validate classes
would chime in before anchor mysql::server::end was reached. This also
caused xinetd and its dependencies to be installed before the mysql
server packages were installed.

This was problematic, because in some cases this would cause package
conflicts, which in turn would cause the mysql server installation to
fail.

The issue was first discovered on Debian 9 with Codership Galera 5.7

To fix this the resource ordering was changed, now galera::status
and galera::validate are executed after anchor mysql::server::end was
reached. According to the (new) tests this does not break anything.