fix hiccup in resource ordering, introduced in #170
Although it is possible to specify the $mysql_service_name class
parameter, internally the module makes an extended evaluation and
hence the class parameter was never intended to be used directly.

So in cases where $mysql_service_name was not specified, chances were
high that an invalid service name was used. This broke the creation
of new setups, because the module tried to create /root/.my.cnf
too early (before the mysql client package was installed).